v0.4.0: Major change: imgui-parameter is removed, now based on ID
As the imgui parameter was pretty verbose, the internal id (m_id) AND a new ID: m_parentId is used to determine the imgui_id. Parent-ID is set automatically when an control is created by a form.

This should make problems due to duplicate IDs go away, if implemented correctly.